parseq/requirements/parseq.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.4
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
joblib==1.3.2
lightning-utilities==0.9.0
multidict==6.0.4
nltk==3.8.1
numpy==1.24.4
packaging==23.1
pillow==10.0.0
pytorch_lightning==1.9.5
pyyaml==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
timm==0.9.5
torchmetrics==1.0.3
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
